implements authentication and authorization page

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,21 @@
 import { useWorkoutsContext } from "../Hooks/useWorkoutsContext";
-import formatDistanceToNow from 'date-fns/formatDistanceToNow'
+import formatDistanceToNow from 'date-fns/formatDistanceToNow';
+import { useAuthContext } from "../Hooks/useAuthContext";
 
 export const WorkoutDetails = ({workout}) => {
 
     const {dispatch} = useWorkoutsContext();
+    const {user} = useAuthContext();
 
     const handleDelete = async() =>{
+        if(!user) {
+            return;
+        }
         const response = await fetch('http://localhost:4024/api/workouts/'+ workout._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
 
         const json = await response.json();
